Move Participant actions out of screeningActions
[#152463922]
