+++ app/javascript/actions/actionTypes.js
@@ -3,12 +3,6 @@ export const CREATE_SCREENING_COMPLETE = 'CREATE_SCREENING_COMPLETE'
 export const FETCH_SCREENING = 'FETCH_SCREENING'
 export const FETCH_SCREENING_COMPLETE = 'FETCH_SCREENING_COMPLETE'
 export const UPDATE_SCREENING_COMPLETE = 'UPDATE_SCREENING_COMPLETE'
-export const CREATE_PARTICIPANT = 'CREATE_PARTICIPANT'
-export const CREATE_PARTICIPANT_COMPLETE = 'CREATE_PARTICIPANT_COMPLETE'
-export const DELETE_PARTICIPANT = 'DELETE_PARTICIPANT'
-export const DELETE_PARTICIPANT_COMPLETE = 'DELETE_PARTICIPANT_COMPLETE'
-export const UPDATE_PARTICIPANT = 'UPDATE_PARTICIPANT'
-export const UPDATE_PARTICIPANT_COMPLETE = 'UPDATE_PARTICIPANT_COMPLETE'
 export const FETCH_HISTORY_OF_INVOLVEMENTS = 'FETCH_HISTORY_OF_INVOLVEMENTS'
 export const FETCH_HISTORY_OF_INVOLVEMENTS_COMPLETE = 'FETCH_HISTORY_OF_INVOLVEMENTS_COMPLETE'
 export const FETCH_RELATIONSHIPS = 'FETCH_RELATIONSHIPS'
