+++ spec/javascripts/actions/personActionsSpec.js
@@ -1,0 +1,15 @@
+import * as actions from 'actions/personActions'
+import {isFSA} from 'flux-standard-action'
+
+describe('personActions', () => {
+  const actionNames = Object.keys(actions)
+  const actionFuncs = actionNames.filter((actionName) => (
+    typeof (actions[actionName]) === 'function'
+  ))
+  actionFuncs.forEach((actionName) => {
+    it(`${actionName} is FSA compliant`, () => {
+      const action = actions[actionName]
+      expect(isFSA(action())).toEqual(true)
+    })
+  })
+})
