+++ spec/javascripts/sagas/deleteParticipantSagaSpec.js
@@ -5,9 +5,10 @@ import {
   deleteParticipantSaga,
   deleteParticipant,
 } from 'sagas/deleteParticipantSaga'
-import {DELETE_PARTICIPANT} from 'actions/actionTypes'
+import {DELETE_PARTICIPANT} from 'actions/personActions'
 import {getScreeningIdValueSelector} from 'selectors/screeningSelectors'
-import * as actions from 'actions/screeningActions'
+import * as personActions from 'actions/personActions'
+import * as screeningActions from 'actions/screeningActions'
 
 describe('deleteParticipantSaga', () => {
   it('deletes participant on DELETE_PARTICIPANT', () => {
@@ -18,23 +19,23 @@ describe('deleteParticipantSaga', () => {
 
 describe('deleteParticipant', () => {
   const id = '123'
-  const action = actions.deleteParticipant(id)
+  const action = personActions.deleteParticipant(id)
 
   it('deletes and puts participant, fetches a screening, and fetches relationships', () => {
     const gen = deleteParticipant(action)
     expect(gen.next().value).toEqual(call(destroy, '/api/v1/participants/123'))
     expect(gen.next().value).toEqual(
-      put(actions.deleteParticipantSuccess(id))
+      put(personActions.deleteParticipantSuccess(id))
     )
     expect(gen.next().value).toEqual(select(getScreeningIdValueSelector))
     expect(gen.next('444').value).toEqual(
-      put(actions.fetchScreening('444'))
+      put(screeningActions.fetchScreening('444'))
     )
     expect(gen.next('444').value).toEqual(
-      put(actions.fetchRelationships('444'))
+      put(screeningActions.fetchRelationships('444'))
     )
     expect(gen.next('444').value).toEqual(
-      put(actions.fetchHistoryOfInvolvements('444'))
+      put(screeningActions.fetchHistoryOfInvolvements('444'))
     )
   })
 
@@ -43,7 +44,7 @@ describe('deleteParticipant', () => {
     const gen = deleteParticipant(action)
     expect(gen.next().value).toEqual(call(destroy, '/api/v1/participants/123'))
     expect(gen.throw(error).value).toEqual(
-      put(actions.deleteParticipantFailure('some error'))
+      put(personActions.deleteParticipantFailure('some error'))
     )
   })
 })
